backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.2.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.4
zope.interface==5.4.0
